+++ Packages/cat.hisa.asset-patcher/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Default language loaded from system if it supports.
+
 ## [1.0.0] - 2024-04-10
 
 - First release.